Crate core_error

Traits for working with Errors.

Usage

If you’re a library, you should reexport all the features exposed by this crate.

Structs

ReportExperimental

An error reporter that prints an error and its sources.

RequestExperimental

Request supports generic, type-driven access to data. Its use is currently restricted to the standard library in cases where trait authors wish to allow trait implementors to share generic information across trait boundaries. The motivating and prototypical use case is core::error::Error which would otherwise require a method per concrete type (eg. std::backtrace::Backtrace instance that implementors want to expose to users).

Traits

Error

Error is a trait representing the basic expectations for error values, i.e., values of type E in Result<T, E>.

Functions

request_refExperimental

Requests a reference of type T from the given impl Error.

request_valueExperimental

Requests a value of type T from the given impl Error.